fix : Hub조회 반환타입 변경... 여기 PR에 넣으면 안되는데... 죄송합니다.

+++ src/main/java/org/gupang/deliveryservice/infrastructure/client/MockHubClient.java
@@ -15,7 +15,7 @@ public class MockHubClient implements HubClient{
     @Override
     public List<HubResponseDto> getHub(UUID StartHubId, UUID endHubId){
         int x = 100;
-        double y = 200d;
+        int y = 200;
         String name1= "name1";
         String name2 = "name2";
         return List.of(new HubResponseDto(
